requirements.txt
flask-appconfig
click
itsdangerous
Jinja2
MarkupSafe
Werkzeug
flask_bootstrap
flask-fontawesome
flask-nav
flask-debug
flask-wtf
pytest
pytest-flask
splinter[flask]
flask-mail